Data storage and scan improvements
Folders are not stored in the DB anymore. Only the individual files. Folders
are implicitly available via the uri/path. This makes the scanning logic much
easier. Now, folder nodes are not inserted on each scan anymore, which makes it
*much* faster on large trees.

It, however, involved an evil and very ugly copy/paste action in the nodes
model. The overall logic should be revised!